requirements.txt
flask
requests
beautifulsoup4
python-telegram-bot==13.15
cloudscraper
gunicorn